A few tweaks to make the lib work well in browsers.
