+++ src/base64.web.ts
@@ -1,0 +1,8 @@
+/**
+ * Encoding a string to base64 in a browser.
+ */
+export function encode(input: string): string {
+
+  return btoa(input);
+
+}
